refactor: split monolithic App.jsx into modular architecture with Tailwind CSS
Replace 1200-line monolith with clean multi-file structure:
- constants/ for theme, film catalog, camera defaults
- utils/ for storage, helpers, film-helpers
- components/ui/ for shadcn-style components (Badge, Button, Input, Select, Card, Sheet, Switch)
- components/ for domain components (FilmRow, StatCard, EmptyState, InfoLine, BarChart, TabBar)
- screens/ for each screen (Dashboard, Stock, AddFilm, FilmDetail, Cameras, Stats)
- App.jsx reduced to routing + state management only

Add Tailwind CSS v4 with @tailwindcss/vite plugin and custom dark theme.
Replace all inline styles with Tailwind utility classes.
