+++ package.json
@@ -17,5 +17,11 @@
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "vite": "^8.0.3"
+  },
+  "devDependencies": {
+    "@tailwindcss/vite": "^4.2.2",
+    "clsx": "^2.1.1",
+    "tailwind-merge": "^3.5.0",
+    "tailwindcss": "^4.2.2"
   }
 }
