+++ src/constants/cameras.js
@@ -1,0 +1,11 @@
+export const DEFAULT_CAMERAS = [
+  { id: "cam1", name: "Canon A-1", format: "35mm", hasInterchangeableBack: false, backs: [] },
+  { id: "cam2", name: "Canon AE-1", format: "35mm", hasInterchangeableBack: false, backs: [] },
+  { id: "cam3", name: "Hasselblad 503cx", format: "120", hasInterchangeableBack: true, backs: [
+    { id: "back1", name: "A12 — Couleur", ref: "A12" },
+    { id: "back2", name: "A12 — N&B", ref: "A12" },
+  ]},
+  { id: "cam4", name: "Yashica Mat 124G", format: "120", hasInterchangeableBack: false, backs: [] },
+  { id: "cam5", name: "Olympus mju-II", format: "35mm", hasInterchangeableBack: false, backs: [] },
+  { id: "cam6", name: "Canon 1000FN", format: "35mm", hasInterchangeableBack: false, backs: [] },
+];
