+++ src/constants/films.js
@@ -1,0 +1,46 @@
+import { Snowflake, Camera, Clock, Eye, Archive } from "lucide-react";
+import { T } from "./theme";
+
+export const FILM_CATALOG = [
+  { id: "portra400", name: "Kodak Portra 400", brand: "Kodak", iso: 400, type: "Couleur", formats: ["35mm","120"] },
+  { id: "portra160", name: "Kodak Portra 160", brand: "Kodak", iso: 160, type: "Couleur", formats: ["35mm","120"] },
+  { id: "portra800", name: "Kodak Portra 800", brand: "Kodak", iso: 800, type: "Couleur", formats: ["35mm","120"] },
+  { id: "gold200", name: "Kodak Gold 200", brand: "Kodak", iso: 200, type: "Couleur", formats: ["35mm"] },
+  { id: "colorplus200", name: "Kodak ColorPlus 200", brand: "Kodak", iso: 200, type: "Couleur", formats: ["35mm"] },
+  { id: "ultramax400", name: "Kodak Ultramax 400", brand: "Kodak", iso: 400, type: "Couleur", formats: ["35mm"] },
+  { id: "ektar100", name: "Kodak Ektar 100", brand: "Kodak", iso: 100, type: "Couleur", formats: ["35mm","120"] },
+  { id: "trix400", name: "Kodak Tri-X 400", brand: "Kodak", iso: 400, type: "N&B", formats: ["35mm","120"] },
+  { id: "tmax400", name: "Kodak T-Max 400", brand: "Kodak", iso: 400, type: "N&B", formats: ["35mm","120"] },
+  { id: "tmax100", name: "Kodak T-Max 100", brand: "Kodak", iso: 100, type: "N&B", formats: ["35mm","120"] },
+  { id: "hp5", name: "Ilford HP5 Plus 400", brand: "Ilford", iso: 400, type: "N&B", formats: ["35mm","120"] },
+  { id: "delta400", name: "Ilford Delta 400", brand: "Ilford", iso: 400, type: "N&B", formats: ["35mm","120"] },
+  { id: "delta3200", name: "Ilford Delta 3200", brand: "Ilford", iso: 3200, type: "N&B", formats: ["35mm","120"] },
+  { id: "fp4", name: "Ilford FP4 Plus 125", brand: "Ilford", iso: 125, type: "N&B", formats: ["35mm","120"] },
+  { id: "panf", name: "Ilford Pan F Plus 50", brand: "Ilford", iso: 50, type: "N&B", formats: ["35mm","120"] },
+  { id: "xp2", name: "Ilford XP2 Super 400", brand: "Ilford", iso: 400, type: "N&B", formats: ["35mm"] },
+  { id: "superia400", name: "Fujifilm Superia X-TRA 400", brand: "Fujifilm", iso: 400, type: "Couleur", formats: ["35mm"] },
+  { id: "c200", name: "Fujifilm C200", brand: "Fujifilm", iso: 200, type: "Couleur", formats: ["35mm"] },
+  { id: "provia100f", name: "Fujifilm Provia 100F", brand: "Fujifilm", iso: 100, type: "Diapo", formats: ["35mm","120"] },
+  { id: "velvia50", name: "Fujifilm Velvia 50", brand: "Fujifilm", iso: 50, type: "Diapo", formats: ["35mm","120"] },
+  { id: "velvia100", name: "Fujifilm Velvia 100", brand: "Fujifilm", iso: 100, type: "Diapo", formats: ["35mm","120"] },
+  { id: "acros100ii", name: "Fujifilm Acros 100 II", brand: "Fujifilm", iso: 100, type: "N&B", formats: ["35mm","120"] },
+  { id: "instaxmini", name: "Fujifilm Instax Mini", brand: "Fujifilm", iso: 800, type: "Instant", formats: ["Instant"] },
+  { id: "instaxwide", name: "Fujifilm Instax Wide", brand: "Fujifilm", iso: 800, type: "Instant", formats: ["Instant"] },
+  { id: "ektachrome", name: "Kodak Ektachrome E100", brand: "Kodak", iso: 100, type: "Diapo", formats: ["35mm","120"] },
+  { id: "cinestill800", name: "CineStill 800T", brand: "CineStill", iso: 800, type: "Couleur", formats: ["35mm","120"] },
+  { id: "cinestill400", name: "CineStill 400D", brand: "CineStill", iso: 400, type: "Couleur", formats: ["35mm","120"] },
+  { id: "cinestill50", name: "CineStill 50D", brand: "CineStill", iso: 50, type: "Couleur", formats: ["35mm"] },
+  { id: "lomo400", name: "Lomography Color Negative 400", brand: "Lomography", iso: 400, type: "Couleur", formats: ["35mm","120"] },
+  { id: "lomo800", name: "Lomography Color Negative 800", brand: "Lomography", iso: 800, type: "Couleur", formats: ["35mm"] },
+  { id: "kentmere400", name: "Kentmere 400", brand: "Kentmere", iso: 400, type: "N&B", formats: ["35mm"] },
+  { id: "fomapan400", name: "Fomapan 400", brand: "Foma", iso: 400, type: "N&B", formats: ["35mm","120"] },
+  { id: "fomapan100", name: "Fomapan 100", brand: "Foma", iso: 100, type: "N&B", formats: ["35mm","120"] },
+];
+
+export const STATES = {
+  stock: { label: "En stock", color: T.blue, icon: Snowflake },
+  loaded: { label: "Chargée", color: T.green, icon: Camera },
+  partial: { label: "Partielle", color: T.amber, icon: Clock },
+  exposed: { label: "Exposée", color: T.accent, icon: Eye },
+  developed: { label: "Développée", color: T.textSec, icon: Archive },
+};
