+++ src/utils/helpers.js
@@ -1,0 +1,6 @@
+export const uid = () => Date.now().toString(36) + Math.random().toString(36).slice(2, 7);
+
+export const today = () => new Date().toISOString().split("T")[0];
+
+export const fmtDate = (d) =>
+  d ? new Date(d).toLocaleDateString("fr-FR", { day: "numeric", month: "short", year: "numeric" }) : "";
